+++ Core/TcpClient.KeepAlive.cs
@@ -1,5 +1,8 @@
 using Dnbn.Models;
 using Microsoft.Extensions.Logging;
+#if NETSTANDARD2_0
+using TaskCompletionSource = Dnbn.Core.TaskCompletionSourceCompat;
+#endif
 
 namespace Dnbn.Core;