+++ Properties/AssemblyInfo.cs
@@ -2,3 +2,4 @@ using System.Runtime.CompilerServices;
 
 // テストプロジェクトからinternalメンバーへのアクセスを許可
 [assembly: InternalsVisibleTo("Dnbn.Tests")]
+[assembly: InternalsVisibleTo("Dnbn.Tests.NetStandard")]