+++ Samples/Dnbn.Sample/Scenarios/Scenario03_Resilience.cs
@@ -3,7 +3,7 @@ using Dnbn.Core;
 using Microsoft.Extensions.Logging;
 using TcpClient = Dnbn.Core.TcpClient;
 
-namespace TcpMessenger.Sample.Scenarios;
+namespace Dnbn.Sample.Scenarios;
 
 /// <summary>
 /// シナリオ3: 障害と自動再接続