+++ Samples/Dnbn.Sample/Scenarios/Scenario05_LegacyProtocols.cs
@@ -6,7 +6,7 @@ using Dnbn.Models;
 using Microsoft.Extensions.Logging;
 using TcpClient = Dnbn.Core.TcpClient;
 
-namespace TcpMessenger.Sample.Scenarios;
+namespace Dnbn.Sample.Scenarios;
 
 /// <summary>
 /// シナリオ5: レガシープロトコル