+++ Samples/Dnbn.Sample/Scenarios/Scenario06_RequestControl.cs
@@ -4,7 +4,7 @@ using Dnbn.Core;
 using Microsoft.Extensions.Logging;
 using TcpClient = Dnbn.Core.TcpClient;
 
-namespace TcpMessenger.Sample.Scenarios;
+namespace Dnbn.Sample.Scenarios;
 
 /// <summary>
 /// シナリオ6: リクエスト制御