+++ Samples/Dnbn.Sample/Scenarios/Scenario07_Monitoring.cs
@@ -7,7 +7,7 @@ using Dnbn.WebUI;
 using Microsoft.Extensions.Logging;
 using TcpClient = Dnbn.Core.TcpClient;
 
-namespace TcpMessenger.Sample.Scenarios;
+namespace Dnbn.Sample.Scenarios;
 
 /// <summary>
 /// シナリオ7: 運用監視